Move FontData.ts to common directory
